fix: remove comment from JSON file to prevent web app crashes
The # CHANGEME comment in the config.json writer is incorrectly included in the output, which is not ignored by Nix. As a result, the web application crashes when trying to load the page.

Although this comment is meant for reference, some users may overlook removing it, leading to the deployment of a broken web application.